backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9.10
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
Pillow>=10.2.0
djangorestframework-simplejwt>=5.3.1
